+++ rollup.test.config.mjs
@@ -20,6 +20,23 @@ export default [
         ],
     },
     {
+        input: 'crates/vertigo/src/driver_module/src_js/api/command/dom/dom_wire.test.ts',
+        output: [
+            {
+                sourcemap: true,
+                file: 'build/dom_wire.test.js',
+                format: 'cjs',
+            }
+        ],
+        plugins: [
+            typescript({
+                sourceMap: true,
+                inlineSources: true,
+            }),
+            sourcemaps(),
+        ],
+    },
+    {
         input: 'crates/vertigo/src/driver_module/src_js/api/command/fetchExec.test.ts',
         output: [
             {